requirements.txt
transformers
torch
pandas
opencv-python-headless
xlsxwriter
scikit-image
imutils
pytesseract
Pillow
